FIX: Recent changes

+++ hummingbot/connector/exchange/idex/types/rest/response.py
@@ -173,7 +173,7 @@ class RestResponseTicker:
 
 @dataclass
 class RestResponseTime:
-    serverTime: float
+    serverTime: int
 
 
 @dataclass
